refactors + fix generate
