+++ packages/cli/src/imports.js
@@ -1,0 +1,3 @@
+export const builder = () => import('@nuxt/builder')
+export const generator = () => import('@nuxt/generator')
+export const core = () => import('@nuxt/core')
