+++ packages/cli/src/index.js
@@ -1,6 +1,8 @@
 import * as _commands from './commands'
+import * as _imports from './imports'
 
 export const commands = _commands
+export const imports = _imports
 
 export { default as setup } from './setup'
 export { default as run } from './run'
